Move some tasks to shared_tasks
